requirements.txt
absl-py==1.0.0
adam==0.0.0.dev0
aiofiles==0.8.0
altair==4.1.0
altgraph==0.17
anaconda==0.0.1.1
anyio==3.5.0
APScheduler==3.6.3
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Augmentor==0.2.9
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.1
blinker==1.4
blis==0.7.4
boltons==21.0.0
boto3==1.20.24
botocore==1.23.24
bulma==1.0.1
cachelib==0.3.0
cachetools==4.2.2
camera==1.3.0
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.2
ci-info==0.2.0
clang==5.0
click==7.1.2
cmake==3.21.2
colorama==0.4.4
compreface-sdk==0.6.0
comtypes==1.1.10
configobj==5.0.6
configparser==5.2.0
constantly==15.1.0
construct==2.10.67
contrib==0.3.0
coremltools==5.0
cryptography==3.4.8
cs50==7.0.2
cssselect==1.1.0
cvzone==1.4.1
cycler==0.10.0
cymem==2.0.5
Cython==0.29.24
data==0.4
debugpy==1.3.0
decorator==4.4.2
deep==0.10
deepface==0.0.70
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.7
dlib==19.22.1
docopt==0.6.2
docutils==0.18.1
download-and-extract==0.3.3
eager==0.0.1
entrypoints==0.3
etelemetry==0.3.0
face==20.1.1
face-compare==1.1.2
Face-rec==0.15
face-recognition==1.3.0
face-recognition-models==0.3.0
face-utils==0.1.0
facenet==1.0.5
facenet-pytorch==2.5.2
fastapi==0.75.2
feedfinder2==0.0.4
feedparser==6.0.8
ffmpeg==1.4
filelock==3.4.0
fitz==0.0.1.dev2
fiz==0.1
Flask==2.0.1
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flatbuffers==1.12
frontend==0.0.3
funcsigs==1.0.2
future==0.18.2
gast==0.4.0
gdown==4.2.0
git-python==1.0.3
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.34.1
gunicorn==20.1.0
h11==0.13.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
httplib2==0.20.4
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
imageai @ https://github.com/OlafenwaMoses/ImageAI/releases/download/2.0.1/imageai-2.0.1-py3-none-any.whl
imageio==2.9.0
imageio-ffmpeg==0.4.5
imutils==0.5.4
incremental==21.3.0
inexactsearch==1.0.1
intel_tensorflow==2.7.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.1
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
jaraco.context==4.0.0
jedi==0.18.0
jieba3k==0.35.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
Js2Py==0.71
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.8.0
Keras-Applications==1.0.8
keras-metrics==1.1.0
keras-models==0.0.7
keras-nightly==2.9.0.dev2022011108
Keras-Preprocessing==1.1.2
keras-vggface==0.6
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lib==3.0.0
libclang==11.1.0
loading==0.0.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mediapipe==0.8.7
mistune==0.8.4
mlxtend==0.19.0
monitoring==0.9.12
more-itertools==8.8.0
moviepy==1.0.3
mpmath==1.2.1
mtcnn==0.1.1
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
newspaper3k==0.2.8
nibabel==3.2.2
nipype==1.7.0
nltk==3.6.2
notebook==6.4.0
numpy==1.22.3
numpy-camera==0.0.2
oauthlib==3.1.1
object-detection-core==0.1
opencv-camera==0.10.11
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.5
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
passlib==1.7.4
pathlib==1.0.1
pathy==0.6.0
pdf2image==1.16.0
pefile==2021.5.24
pickleshare==0.7.5
Pillow==8.3.1
pipwin==0.5.1
platformdirs==2.3.0
ply==3.11
preshed==3.0.5
priority==1.3.0
proglog==0.1.9
prometheus-client==0.11.0
prompt-toolkit==3.0.19
Protego==0.1.16
protobuf==3.17.3
protos==0.1.0
prov==2.0.0
psutil==5.9.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
pydantic==1.8.2
pydeck==0.6.2
PyDispatcher==2.0.5
pydot==1.4.2
pygame==2.0.1
Pygments==2.9.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyjsparser==2.7.1
PyJWT==2.3.0
PyMuPDF==1.19.6
PyMySQL==1.0.2
pynput==1.7.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
pyrsistent==0.18.0
PySimpleGUI==4.47.0
pySmartDL==1.3.4
pysmi==0.3.4
pysnmp==4.4.12
PySocks==1.7.1
pyspellchecker==0.6.2
pytesseract==0.3.8
python-dateutil==2.8.2
python-multipart==0.0.5
python-telegram-bot==13.7
python-time==0.3.0
pytils==0.4.1
pytime==0.2.2
pyttsx==1.1
pyttsx3==2.90
pytz==2021.1
pyvirtualcam==0.9.0
PyWavelets==1.1.1
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.3
pyxnat==1.4
PyYAML==5.4.1
pyzmq==22.1.0
queuelib==1.6.2
rdflib==6.1.1
regex==2021.8.28
request-id==1.0.1
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retina-face==0.0.7
rsa==4.7.2
s3transfer==0.5.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.8.0
Scrapy==2.5.0
seaborn==0.11.2
Send2Trash==1.7.1
service-identity==21.1.0
sgmllib3k==1.0.0
silpa-common==0.3
simplejson==3.17.5
six==1.15.0
slurm==0.4.6
smart-open==5.2.1
smmap==4.0.0
sniffio==1.2.0
soundex==1.1.3
soupsieve==2.2.1
spacy==3.1.3
spacy-legacy==3.0.8
SpeechRecognition==3.8.1
spellchecker==0.4
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.2
srsly==2.4.1
starlette==0.17.1
streamlit==0.84.2
sympy==1.9
tb-nightly==2.8.0a20220111
telegram==0.0.1
tensor==0.3.6
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.6.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.10.1
tesseract==0.1.3
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
tf-nightly==2.9.0.dev20220111
thinc==8.0.10
threadpoolctl==2.2.0
tifffile==2021.11.2
tinysegmenter==0.3
tldextract==3.1.1
toml==0.10.2
tools==0.1.9
toolz==0.11.1
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
traits==6.3.2
twilio==7.4.0
Twisted==21.7.0
twisted-iocpsupport==1.0.1
typer==0.4.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.6
utils==1.0.1
uvicorn==0.17.6
validators==0.18.2
virtualenv==20.10.0
w3lib==1.22.0
wasabi==0.8.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wincertstore==0.2
wolframalpha==5.0.0
wrapt==1.12.1
WTForms==2.3.3
xmltodict==0.12.0
zope.interface==5.4.0
